alert to confirm the password when deleting

+++ lib/pages/profile_page.dart
@@ -8,7 +8,7 @@ import 'package:firebase_storage/firebase_storage.dart' as firebase_storage;
 import '../components/text_box.dart';
 
 class ProfilePage extends StatefulWidget {
-  const ProfilePage({super.key});
+  const ProfilePage({Key? key}) : super(key: key);
 
   @override
   State<ProfilePage> createState() => _ProfilePageState();
@@ -57,33 +57,83 @@ class _ProfilePageState extends State<ProfilePage> {
     }
   }
 
+  Future<void> reauthenticateUser(String password) async {
+    UserCredential credential =
+        await FirebaseAuth.instance.signInWithEmailAndPassword(
+      email: currentUser.email!,
+      password: password,
+    );
+  }
+
   Future<void> deleteProfile() async {
-    final bool confirmDelete = await showDialog(
+    String password = "";
+
+    await showDialog(
       context: context,
       builder: (context) => AlertDialog(
         backgroundColor: Colors.grey[900],
-        title:
-            const Text("Delete Profile", style: TextStyle(color: Colors.white)),
-        content: const Text("Are you sure you want to delete your profile?",
-            style: TextStyle(color: Colors.white)),
+        title: Text("Enter Password", style: TextStyle(color: Colors.white)),
+        content: TextField(
+          obscureText: true,
+          style: TextStyle(color: Colors.white),
+          decoration: InputDecoration(
+            hintText: "Enter your password",
+            hintStyle: TextStyle(color: Colors.grey),
+          ),
+          onChanged: (value) {
+            password = value;
+          },
+        ),
         actions: [
           TextButton(
-            child: const Text('Cancel', style: TextStyle(color: Colors.white)),
-            onPressed: () => Navigator.pop(context, false),
+            child: Text('Cancel', style: TextStyle(color: Colors.white)),
+            onPressed: () => Navigator.pop(context),
           ),
           TextButton(
-            child: const Text('Delete', style: TextStyle(color: Colors.red)),
-            onPressed: () => Navigator.pop(context, true),
+            child: Text('Submit', style: TextStyle(color: Colors.white)),
+            onPressed: () async {
+              try {
+                await reauthenticateUser(password);
+                Navigator.pop(context, true);
+              } catch (e) {
+                // Handle reauthentication failure (e.g., incorrect password)
+                print("Reauthentication failed: $e");
+              }
+            },
           )
         ],
       ),
     );
 
-    if (confirmDelete == true) {
-      await usersCollection.doc(currentUser.email).delete();
-      await currentUser.delete();
-      await FirebaseAuth.instance.signOut();
-      Navigator.pop(context);
+    if (password.isNotEmpty) {
+      final bool confirmDelete = await showDialog(
+        context: context,
+        builder: (context) => AlertDialog(
+          backgroundColor: Colors.grey[900],
+          title: const Text("Delete Profile",
+              style: TextStyle(color: Colors.white)),
+          content: const Text("Are you sure you want to delete your profile?",
+              style: TextStyle(color: Colors.white)),
+          actions: [
+            TextButton(
+              child:
+                  const Text('Cancel', style: TextStyle(color: Colors.white)),
+              onPressed: () => Navigator.pop(context, false),
+            ),
+            TextButton(
+              child: const Text('Delete', style: TextStyle(color: Colors.red)),
+              onPressed: () => Navigator.pop(context, true),
+            )
+          ],
+        ),
+      );
+
+      if (confirmDelete == true) {
+        await usersCollection.doc(currentUser.email).delete();
+        await currentUser.delete();
+        await FirebaseAuth.instance.signOut();
+        Navigator.pop(context);
+      }
     }
   }
